requirements.txt
flask==2.0.1
selenium==4.1.0
beautifulsoup4==4.9.3
requests==2.26.0
pillow==8.3.2
pandas==1.3.3
numpy==1.21.2
tqdm==4.62.3
webdriver_manager==3.5.2